refactor(dot11): split constant and label tests into tests codepoints module
